requirements.txt
attrs==26.1.0
boto3==1.42.76
botocore==1.42.76
certifi==2026.2.25
cffi==2.0.0
chromedriver-autoinstaller==0.6.4
h11==0.16.0
html5lib==1.1
idna==3.11
jmespath==1.1.0
numpy==2.4.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pycparser==3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
s3transfer==0.16.0
selenium==4.27.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websocket-client==1.9.0
wheel==0.46.3
wsproto==1.3.2
xlrd==2.0.1
